// src/utils/zhihu-billboard.d.ts
/* eslint-disable */

export interface RecordHot {
  /** 标题 */
  title: string;
  /** 描述 */
  description: string;
  /** 热度指数 */
  metrics: string;
  /** 图片信息 */
  img: string;
  /** 跳转链接 */
  url: string;
}

export interface RecordHotRaw {
  "type": "hot_list_feed",
  "styleType": "1",
  "id": "0_1609768003.2811854",
  "cardId": "Q_437783708",
  "feedSpecific": {
    "answerCount": 3417
  },
  "target": {
    "titleArea": {
      "text": "拼多多疑似回应「员工猝死」后秒删，是真的吗？拼多多又称截图系谣言，你相信吗？"
    },
    "excerptArea": {
      "text": "知乎官方回应：「拼多多」系知乎注册用户，其身份真实无误。知乎有严格的身份认证流程和机制。 4 日 8 时 19 分 49 秒，「拼多多」创建回答「你们看看底层的人民，哪一个不是用命换钱，我一直不以为是资本的问题，而是这个社会的问题，这是一个用命拼的时代，你可以选择安逸的日子，但你就要选择安逸带来的后果，人是可以控制自己的努力的，我们都可以」。 4 日 8 时 20 分 17 秒，「拼多多」自行删除了上述回答。 拼多多最新回应：此前「员工猝死」的发文截图不实，你相信吗？"
    },
    "imageArea": {
      "url": "https:\u002F\u002Fpic4.zhimg.com\u002F80\u002Fv2-7bc01cb7a58e5db84bab89be95dfc9ce_720w.png"
    },
    "metricsArea": {
      "text": "24642 万热度"
    },
    "labelArea": {
      "type": "trend",
      "trend": 0,
      "nightColor": "#B7302D",
      "normalColor": "#F1403C"
    },
    "link": {
      "url": "https:\u002F\u002Fwww.zhihu.com\u002Fquestion\u002F437783708"
    },
    "textTagArea": {
      "text": "知乎官方回应",
      "fontColor": "GYL01A",
      "background": "GYL01A"
    }
  },
  "attachedInfo": "CjgIABADGgg1OTU2NDAzMSDNgMr\u002FBTC+ATj2UUAAcgk0Mzc3ODM3MDh4AKoBCWJpbGxib2FyZNIBAA=="
}